vector utilities
